feat(web): regroup sidebar + consolidate scan entry points (IA restructure ①②)
Addresses 'buttons scattered, no pattern' and 'newcomer has no entry point'
by restructuring the information architecture — the real UX problem, not
the earlier code-hardening passes.

① Sidebar regroup (11 flat items → 3 sections):
- Core: Dashboard / Devices / Changes
- Library: Documents
- Administration (admin-only, hidden for non-admins): Networks / Agents /
  Users / Audit
Previously the 4 admin-only items rendered as inert grey rows mixed in with
daily-use items, reading as 'a bunch of things I can't click'. Non-admins now
see 4 items + footer instead of 11. /topology and /discovery removed as
top-level entries (topology → view toggle in /devices; discovery → scan tab).
Added nav-group-label styling + dividers.

② Scan entry consolidation (4 scattered → 1 scan center under Devices):
- /devices sub-nav (already had Scanner/Scan Tasks/Scan Results tabs) extended
  with a 'Passive Discovery' tab → new /devices/discovery route (content moved
  from standalone /discovery, which now redirects).
- Tab order reorganized daily→occasional: Scan Tasks / Scan Results / Quick
  Scan (was 'Scanner', vague) / Passive Discovery.
- Deleted /scanner and /scanner/results redirect stubs (pure goto, zero
  inbound links). /discovery now redirects to /devices/discovery.

Net: every 'find devices on the network' path now enters through Devices →
scan tabs; no more guessing between /discovery vs /scanner vs /devices/scanner.
