+++ web/src/routes/+layout.svelte
@@ -32,9 +32,7 @@
 		X,
 		History,
 		Bot,
-		Radar,
-		Network,
-		Share2
+		Network
 	} from '@lucide/svelte';
 	import type { Component } from 'svelte';
 
@@ -56,21 +54,46 @@
 		goto('/login');
 	}
 
-	const allNavItems: { path: string; label: string; icon: Component; adminOnly?: boolean }[] = [
-		{ path: '/dashboard', label: m['navigation.Dashboard'](), icon: LayoutDashboard },
-		{ path: '/devices', label: m['navigation.Devices'](), icon: Server },
-		{ path: '/changes', label: m['navigation.Changes'](), icon: History },
-		{ path: '/topology', label: m['navigation.Topology'](), icon: Share2 },
-		{ path: '/discovery', label: m['navigation.Discovery'](), icon: Radar },
-		{ path: '/networks', label: m['navigation.Networks'](), icon: Network, adminOnly: true },
-		{ path: '/agents', label: m['navigation.Agents'](), icon: Bot, adminOnly: true },
-		{ path: '/documents', label: m['navigation.Documents'](), icon: FileText },
-		{ path: '/users', label: m['navigation.Users'](), icon: Users, adminOnly: true },
-		{ path: '/audit', label: m['navigation.AuditLogs'](), icon: ScrollText, adminOnly: true },
-		{ path: '/settings', label: m['navigation.Settings'](), icon: Settings }
+	// Nav is grouped (Core / Library / Administration) instead of a flat 11-item
+	// list. The admin group is hidden entirely for non-admins — previously the
+	// 4 admin-only items rendered as inert grey rows mixed in with the daily-use
+	// items, which read as "a bunch of things I can't click". /topology and
+	// /discovery were removed as top-level entries: topology is now a view toggle
+	// inside /devices, discovery is a tab under the scan center (/devices/scan-*).
+	interface NavItem { path: string; label: string; icon: Component }
+	interface NavGroup { labelKey: string; items: NavItem[] }
+
+	const navGroups: NavGroup[] = [
+		{
+			labelKey: 'navigation.Group Core',
+			items: [
+				{ path: '/dashboard', label: m['navigation.Dashboard'](), icon: LayoutDashboard },
+				{ path: '/devices', label: m['navigation.Devices'](), icon: Server },
+				{ path: '/changes', label: m['navigation.Changes'](), icon: History }
+			]
+		},
+		{
+			labelKey: 'navigation.Group Library',
+			items: [
+				{ path: '/documents', label: m['navigation.Documents'](), icon: FileText }
+			]
+		},
+		{
+			labelKey: 'navigation.Group Admin',
+			items: [
+				{ path: '/networks', label: m['navigation.Networks'](), icon: Network },
+				{ path: '/agents', label: m['navigation.Agents'](), icon: Bot },
+				{ path: '/users', label: m['navigation.Users'](), icon: Users },
+				{ path: '/audit', label: m['navigation.AuditLogs'](), icon: ScrollText }
+			]
+		}
 	];
 
-	const navItems = $derived(allNavItems.filter(item => !item.adminOnly || $auth.user?.role === 'admin'));
+	// Non-admins don't see the Administration group at all. The Settings entry
+	// stays in the footer (it's self-service profile/security, not admin infra).
+	const visibleGroups = $derived(
+		navGroups.filter(g => g.labelKey !== 'navigation.Group Admin' || $auth.user?.role === 'admin')
+	);
 
 	// Global keyboard shortcuts
 	$effect(() => {
@@ -157,20 +180,26 @@
 				</div>
 			</div>
 		</div>
-		<nav class="flex-1 p-2">
-			{#each navItems as item (item.path)}
-				{@const active = $page.url.pathname === item.path || (item.path !== '/' && $page.url.pathname.startsWith(item.path))}
-				<a
-					href={item.path}
-					class="flex items-center gap-3 px-3 py-2.5 rounded-md mb-0.5 text-sm transition-colors
-						{active
-							? 'bg-primary/10 text-primary font-medium border-l-2 border-primary'
-							: 'text-muted hover:bg-surface-2 hover:text-text border-l-2 border-transparent'}"
-					aria-current={active ? 'page' : undefined}
-				>
-					<item.icon class="w-[18px] h-[18px] shrink-0" />
-					<span>{item.label}</span>
-				</a>
+		<nav class="flex-1 p-2 overflow-y-auto">
+			{#each visibleGroups as group, gi (group.labelKey)}
+				{#if gi > 0}
+					<div class="my-2 mx-3 border-t border-border"></div>
+				{/if}
+				<p class="nav-group-label">{m[group.labelKey]()}</p>
+				{#each group.items as item (item.path)}
+					{@const active = $page.url.pathname === item.path || (item.path !== '/' && $page.url.pathname.startsWith(item.path))}
+					<a
+						href={item.path}
+						class="flex items-center gap-3 px-3 py-2.5 rounded-md mb-0.5 text-sm transition-colors
+							{active
+								? 'bg-primary/10 text-primary font-medium border-l-2 border-primary'
+								: 'text-muted hover:bg-surface-2 hover:text-text border-l-2 border-transparent'}"
+						aria-current={active ? 'page' : undefined}
+					>
+						<item.icon class="w-[18px] h-[18px] shrink-0" />
+						<span>{item.label}</span>
+					</a>
+				{/each}
 			{/each}
 		</nav>
 		<div class="p-3 border-t border-border space-y-2">
@@ -212,3 +241,22 @@
 </div>
 
 <Toast />
+
+<style>
+	/* Section label for grouped nav (Core / Library / Administration).
+	 * Visually subdued so the items remain the focus, but present so the
+	 * grouping is legible — especially the admin/normal split. */
+	.nav-group-label {
+		padding: 0.625rem 0.75rem 0.25rem;
+		font-size: 0.6875rem;
+		font-weight: 600;
+		text-transform: uppercase;
+		letter-spacing: 0.05em;
+		color: var(--color-text-muted);
+		opacity: 0.7;
+	}
+	/* First group has no top label padding (the header above already has padding). */
+	.nav-group-label:first-child {
+		padding-top: 0.25rem;
+	}
+</style>
