--- web/src/routes/scanner/+page.svelte
@@ -1,18 +1,0 @@
-<!--
-  SPDX-License-Identifier: AGPL-3.0-or-later
-
-  Copyright (c) 2026 Mi-Bee Studio. All rights reserved.
-
-  This file is part of MiBee Steward, distributed under the GNU Affero General
-  Public License v3.0 or later. A commercial license is available for use cases
-  the AGPL does not accommodate; see LICENSE-COMMERCIAL.md.
--->
-
-<script lang="ts">
-	import { goto } from '$app/navigation';
-	import { onMount } from 'svelte';
-
-	onMount(() => {
-		goto('/devices/scan-tasks', { replaceState: true });
-	});
-</script>
